🎨 실시간 채팅하는 공간을 component로 분리
