+++ client/src/pages/Chatting/views/ChattingPage.tsx
@@ -1,69 +1,13 @@
-import { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
-import { Client } from '@stomp/stompjs';
-import SockJS from 'sockjs-client';
+import ChatRoomList from '../components/ChatRoomList';
+import ChattingArea from '../components/ChattingArea';
 
 function ChattingPage() {
-  const { roomId } = useParams<{ roomId: string }>();
-  const [client, setClient] = useState<Client | null>(null);
-  const [messages, setMessages] = useState<string[]>([]);
-  const [newMessage, setNewMessage] = useState<string>('');
-
-  useEffect(() => {
-    const newClient = new Client({
-      webSocketFactory: () =>
-        new SockJS(process.env.REACT_APP_API_URL + '/chat'),
-    });
-    setClient(newClient);
-  }, []);
-
-  console.log('roomId', roomId);
-  console.log('client', client);
-
-  useEffect(() => {
-    if (client) {
-      client.onConnect = function (frame) {
-        client.subscribe(`/topic/${roomId}`, function (message) {
-          setMessages((prev) => [...prev, message.body]);
-        });
-      };
-      client.onStompError = function (frame) {
-        console.log('Broker reported error: ' + frame.headers['message']);
-        console.log('Additional details: ' + frame.body);
-      };
-      client.activate();
-    }
-
-    return () => {
-      if (client) {
-        client.deactivate();
-      }
-    };
-  }, [client, roomId]);
-
-  const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setNewMessage(event.target.value);
-    console.log('newMessage', newMessage);
-  };
-
-  const onSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (client) {
-      client.publish({
-        destination: `/app/chat/${roomId}`,
-        body: newMessage,
-      });
-    }
-    setNewMessage('');
-  };
-
   return (
     <>
-      <div>chatting page</div>
-      <form onSubmit={onSubmit}>
-        <input value={newMessage} onChange={onChange} />
-        <button>전송</button>
-      </form>
+      // ChatRoomList는 채팅방 목록을 보여줍니다.
+      <ChatRoomList />
+      // ChattingArea는 ChatRoomList에서 선택한 채팅방의 roomId를 사용합니다.
+      <ChattingArea />
     </>
   );
 }
